[TASK] #maintenance - added dynamic imports to TypeScript structure
