+++ assets/ts/controllerMapper.ts
@@ -1,0 +1,9 @@
+//import { PinController } from './controller/PinController';
+import { CalendarController } from "./controller/CalendarController";
+
+export const controllerMapper: Record<string, any> = {
+//    pin: PinController
+    calendar: CalendarController
+};
+
+// default?
