+++ package.json
@@ -1,6 +1,7 @@
 {
     "devDependencies": {
         "@babel/core": "^7.17.0",
+        "@babel/plugin-syntax-dynamic-import": "^7.8.3",
         "@babel/preset-env": "^7.16.0",
         "@symfony/webpack-encore": "^5.0.0",
         "core-js": "^3.38.0",
