+++ templates/board.html.twig
@@ -1,6 +1,7 @@
 {% extends "base.html.twig" %}
 
 {% block title %}Board{% endblock %}
+{% block pageType %}homepage{% endblock %}
 
 {% block body %}
     <h1>Board</h1>
