+++ templates/calendar.html.twig
@@ -1,6 +1,7 @@
 {% extends "base.html.twig" %}
 
 {% block title %}Kalender{% endblock %}
+{% block pageType %}calendar{% endblock %}
 
 {% block body %}
     <div id="calendar-holder"></div>
